[TUBEMQ-503] offset query + design by contract
