+++ tubemq-manager/src/main/java/org/apache/tubemq/manager/controller/group/result/AllBrokersOffsetRes.java
@@ -15,29 +15,21 @@
  * limitations under the License.
  */
 
+package org.apache.tubemq.manager.controller.group.result;
 
-package org.apache.tubemq.manager.service.tube;
+import java.util.ArrayList;
+import java.util.List;
+import lombok.Data;
 
-public class TubeHttpBrokerCfgInfo {
-    private boolean acceptPublish;
-    private boolean acceptSubscribe;
-    private Integer brokerId;
-    private String brokerIp;
-    private Integer brokerPort;
-    private Integer brokerTLSPort;
-    private String createDate;
-    private String createUser;
-    private String deletePolicy;
-    private String deleteWhen;
-    private String modifyDate;
-    private String modifyUser;
-    private Integer memCacheFlushIntvl;
-    private Integer memCacheMsgCntInK;
-    private Integer memCacheMsgSizeInMB;
-    private Integer numPartitions;
-    private Integer numTopicStores;
-    private Integer unflushDataHold;
-    private Integer unflushInterval;
-    private Integer unflushThreshold;
-    private boolean hasTLSPort;
+
+@Data
+public class AllBrokersOffsetRes {
+
+    private List<OffsetInfo> offsetPerBroker = new ArrayList<>();
+
+    @Data
+    public static class OffsetInfo {
+        private int brokerId;
+        private OffsetQueryRes offsetQueryRes;
+    }
 }
