+++ tubemq-manager/src/main/resources/application.properties
